pyproject.toml
[project]
name = "confluence-to-bookstack"
version = "0.1.0"
description = ""
authors = [
    {name = "Berrah Mohamed"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "argparse (>=1.4.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "pydantic (>=2.0.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false